requirements.txt
discord.py==1.7.3
Django>=3.2.13
emojis==0.6.0
GitPython==3.1.27
praw==7.0.0
ratelimit==2.2.1
requests==2.23.0
schedule==0.6.0
pytest==7.1.1
pytest-cov==3.0.0
flake8==4.0.1
black==22.3.0
